+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 1.2.0
+
+### Features
+- Added `rnd_eject` which ejects a random person from the ship
+
+
+### Minor
+- Changed `eject` command and `list` to iterate using foreach loop instead of regular for-loop
+
 ## 1.1.0
 
 - Messages can now be more than just 1 word